+++ server/TWS Foundation/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## CURENT - [xx.xx-xxxx]
 
-- Notes:
-
+- Notes: 
+    1. Implemented [Carriers] and [VehiculesModels] service.
+    2. [UpdateHelper] method, in Update repository now add new items to a icollection
+    if the item is provided.
+    
 - Dependencies Upgrade: N/A
 
 ## 4.0.0 - [24.10-2024]